python_worker/requirements.txt
# Core Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
python-dotenv==1.0.1

# Database & Event System
supabase==2.5.1

# AI/ML Libraries
google-genai>=1.65.0
sentence-transformers==3.0.1
scikit-learn==1.5.1
scikit-fuzzy==0.4.2
numpy==1.26.4
scipy==1.13.1
torch==2.3.1

# Document Processing
PyMuPDF==1.24.5
python-docx==1.1.2

# Cryptography (badge signing)
cryptography>=42.0.0

# Utilities
httpx==0.27.0
tenacity==8.4.2
